pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pr-writer"
version = "0.1.0"
description = "AI-powered PR description generator using Claude"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyperclip>=1.9.0",
    "requests>=2.31.0",
]

[project.scripts]
pr-writer = "pr_writer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pr_writer*"]
